backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose
passlib
email-validator
python-multipart
resend
psycopg2-binary
sqlalchemy